moved parts of browser to its own file |  #1694
