+++ extension/chrome/dev/storage.ts
@@ -6,6 +6,7 @@ import { Catch } from '../../js/common/platform/catch.js';
 import { Store, Storable, AccountStore, GlobalStore, GlobalIndex, AccountIndex, RawStore } from '../../js/common/platform/store.js';
 import { Str, Dict } from '../../js/common/core/common.js';
 import { Xss, Ui, Env } from '../../js/common/browser.js';
+import { Assert } from '../../js/common/assert.js';
 
 Catch.try(async () => {
 
@@ -14,9 +15,9 @@ Catch.try(async () => {
   const DEBUG_EMAILS = ['info@nvimp.com', 'human@flowcrypt.com', 'flowcrypt.compatibility@gmail.com'];
 
   const uncheckedUrlParams = Env.urlParams(['filter', 'keys', 'controls', 'title']);
-  const filter = Env.urlParamRequire.optionalString(uncheckedUrlParams, 'filter');
-  const keys = Env.urlParamRequire.optionalString(uncheckedUrlParams, 'keys');
-  const title = Env.urlParamRequire.optionalString(uncheckedUrlParams, 'title');
+  const filter = Assert.urlParamRequire.optionalString(uncheckedUrlParams, 'filter');
+  const keys = Assert.urlParamRequire.optionalString(uncheckedUrlParams, 'keys');
+  const title = Assert.urlParamRequire.optionalString(uncheckedUrlParams, 'title');
   const controls = uncheckedUrlParams.controls === true && (Catch.environment().includes(':dev') || DEBUG_EMAILS.includes(String(filter)));
 
   if (title) {
