+++ extension/chrome/elements/add_pubkey.ts
@@ -4,18 +4,21 @@
 
 import { Catch } from '../../js/common/platform/catch.js';
 import { Store } from '../../js/common/platform/store.js';
-import { Xss, Ui, KeyImportUi, AttUI, processPublicKeyFileImport, UserAlert, Env } from '../../js/common/browser.js';
+import { Xss, Ui, Env } from '../../js/common/browser.js';
 import { BrowserMsg } from '../../js/common/extension.js';
+import { Assert } from '../../js/common/assert.js';
+import { KeyImportUi, processPublicKeyFileImport, UserAlert, } from '../../js/common/ui/key_import_ui.js';
+import { AttUI } from '../../js/common/ui/att_ui.js';
 
 Catch.try(async () => {
 
   Ui.event.protect();
 
   const uncheckedUrlParams = Env.urlParams(['acctEmail', 'parentTabId', 'emails', 'placement']);
-  const acctEmail = Env.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
-  const parentTabId = Env.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
-  const missingPubkeyEmails = Env.urlParamRequire.string(uncheckedUrlParams, 'emails').split(',');
-  const placement = Env.urlParamRequire.optionalString(uncheckedUrlParams, 'placement');
+  const acctEmail = Assert.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
+  const parentTabId = Assert.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
+  const missingPubkeyEmails = Assert.urlParamRequire.string(uncheckedUrlParams, 'emails').split(',');
+  const placement = Assert.urlParamRequire.optionalString(uncheckedUrlParams, 'placement');
 
   for (const missingPubkeyEmail of missingPubkeyEmails) {
     Xss.sanitizeAppend('select.email', `<option value="${Xss.escape(missingPubkeyEmail)}">${Xss.escape(missingPubkeyEmail)}</option>`);
