+++ extension/chrome/elements/pgp_pubkey.ts
@@ -9,6 +9,7 @@ import { Xss, Ui, Env } from '../../js/common/browser.js';
 import { mnemonic } from '../../js/common/core/mnemonic.js';
 import { Pgp, Contact } from '../../js/common/core/pgp.js';
 import { BrowserMsg } from '../../js/common/extension.js';
+import { Assert } from '../../js/common/assert.js';
 
 declare const openpgp: typeof OpenPGP;
 
@@ -20,10 +21,10 @@ Catch.try(async () => {
 
   // minimized means I have to click to see details. Compact means the details take up very little space.
   const uncheckedUrlParams = Env.urlParams(['acctEmail', 'armoredPubkey', 'parentTabId', 'minimized', 'compact', 'frameId']);
-  const acctEmail = Env.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
-  const parentTabId = Env.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
-  const armoredPubkey = Env.urlParamRequire.string(uncheckedUrlParams, 'armoredPubkey');
-  const frameId = Env.urlParamRequire.string(uncheckedUrlParams, 'frameId');
+  const acctEmail = Assert.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
+  const parentTabId = Assert.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
+  const armoredPubkey = Assert.urlParamRequire.string(uncheckedUrlParams, 'armoredPubkey');
+  const frameId = Assert.urlParamRequire.string(uncheckedUrlParams, 'frameId');
   const compact = uncheckedUrlParams.compact === true;
   const minimized = uncheckedUrlParams.minimized === true;
