+++ extension/chrome/elements/reply_pubkey_mismatch.ts
@@ -10,18 +10,19 @@ import { Api } from '../../js/common/api/api.js';
 import { BrowserMsg } from '../../js/common/extension.js';
 import { Catch } from '../../js/common/platform/catch.js';
 import { Google } from '../../js/common/api/google.js';
+import { Assert } from '../../js/common/assert.js';
 
 Catch.try(async () => {
 
   Ui.event.protect();
 
   const uncheckedUrlParams = Env.urlParams(['acctEmail', 'from', 'to', 'subject', 'frameId', 'threadId', 'threadMsgId', 'parentTabId', 'skipClickPrompt', 'ignoreDraft', 'debug']);
-  const acctEmail = Env.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
-  const parentTabId = Env.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
-  const from = Env.urlParamRequire.optionalString(uncheckedUrlParams, 'from') || acctEmail;
-  const subject = Env.urlParamRequire.optionalString(uncheckedUrlParams, 'subject') || '';
-  const frameId = Env.urlParamRequire.string(uncheckedUrlParams, 'frameId');
-  const threadId = Env.urlParamRequire.optionalString(uncheckedUrlParams, 'threadId') || '';
+  const acctEmail = Assert.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
+  const parentTabId = Assert.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
+  const from = Assert.urlParamRequire.optionalString(uncheckedUrlParams, 'from') || acctEmail;
+  const subject = Assert.urlParamRequire.optionalString(uncheckedUrlParams, 'subject') || '';
+  const frameId = Assert.urlParamRequire.string(uncheckedUrlParams, 'frameId');
+  const threadId = Assert.urlParamRequire.optionalString(uncheckedUrlParams, 'threadId') || '';
   const to = uncheckedUrlParams.to ? String(uncheckedUrlParams.to).split(',') : [];
   const debug = uncheckedUrlParams.debug === true;
