+++ extension/chrome/elements/sending_address.ts
@@ -10,12 +10,13 @@ import { Settings } from '../../js/common/settings.js';
 import { BrowserMsg } from '../../js/common/extension.js';
 import { Api } from '../../js/common/api/api.js';
 import { Buf } from '../../js/common/core/buf.js';
+import { Assert } from '../../js/common/assert.js';
 
 Catch.try(async () => {
 
   const uncheckedUrlParams = Env.urlParams(['acctEmail', 'parentTabId', 'placement']);
-  const acctEmail = Env.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
-  const parentTabId = Env.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
+  const acctEmail = Assert.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
+  const parentTabId = Assert.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
   const container = $('.emails');
 
   const storage = await Store.getAcct(acctEmail, ['addresses']);
