+++ extension/chrome/elements/shared/footer.ts
@@ -7,15 +7,16 @@ import { Catch } from '../../../js/common/platform/catch.js';
 import { Store } from '../../../js/common/platform/store.js';
 import { Ui, Env } from '../../../js/common/browser.js';
 import { BrowserMsg } from '../../../js/common/extension.js';
+import { Assert } from '../../../js/common/assert.js';
 
 Catch.try(async () => {
 
   Ui.event.protect();
 
   const uncheckedUrlParams = Env.urlParams(['acctEmail', 'parentTabId', 'grandparentTabId']); // placement: compose||settings
-  const acctEmail = Env.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
-  const parentTabId = Env.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
-  const grandparentTabId = Env.urlParamRequire.string(uncheckedUrlParams, 'grandparentTabId');  // grandparent is the email provider tab
+  const acctEmail = Assert.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
+  const parentTabId = Assert.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
+  const grandparentTabId = Assert.urlParamRequire.string(uncheckedUrlParams, 'grandparentTabId');  // grandparent is the email provider tab
 
   const renderInitial = async () => {
     const subscription = await Store.subscription();
