+++ extension/chrome/settings/modules/add_key.ts
@@ -5,17 +5,19 @@
 import { Catch } from '../../../js/common/platform/catch.js';
 import { Store } from '../../../js/common/platform/store.js';
 import { Value } from '../../../js/common/core/common.js';
-import { Xss, Ui, KeyImportUi, UserAlert, KeyCanBeFixed, Env } from '../../../js/common/browser.js';
+import { Xss, Ui, Env } from '../../../js/common/browser.js';
 import { Pgp } from '../../../js/common/core/pgp.js';
 import { Api } from '../../../js/common/api/api.js';
 import { BrowserMsg } from '../../../js/common/extension.js';
 import { Google } from '../../../js/common/api/google.js';
+import { Assert } from '../../../js/common/assert.js';
+import { KeyImportUi, UserAlert, KeyCanBeFixed } from '../../../js/common/ui/key_import_ui.js';
 
 Catch.try(async () => {
 
   const uncheckedUrlParams = Env.urlParams(['acctEmail', 'parentTabId']);
-  const acctEmail = Env.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
-  const parentTabId = Env.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
+  const acctEmail = Assert.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
+  const parentTabId = Assert.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
 
   await Ui.passphraseToggle(['input_passphrase']);
   const keyImportUi = new KeyImportUi({ rejectKnown: true });
