+++ extension/chrome/settings/modules/decrypt.ts
@@ -5,14 +5,17 @@
 import { Catch } from '../../../js/common/platform/catch.js';
 import { Store } from '../../../js/common/platform/store.js';
 import { Att } from '../../../js/common/core/att.js';
-import { Xss, Ui, XssSafeFactory, AttUI, Env, Browser } from '../../../js/common/browser.js';
+import { Xss, Ui, Env, Browser } from '../../../js/common/browser.js';
 import { BrowserMsg } from '../../../js/common/extension.js';
 import { DecryptErrTypes, PgpMsg } from '../../../js/common/core/pgp.js';
+import { Assert } from '../../../js/common/assert.js';
+import { AttUI } from '../../../js/common/ui/att_ui.js';
+import { XssSafeFactory } from '../../../js/common/xss_safe_factory.js';
 
 Catch.try(async () => {
 
   const uncheckedUrlParams = Env.urlParams(['acctEmail', 'parentTabId']);
-  const acctEmail = Env.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
+  const acctEmail = Assert.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
 
   const tabId = await BrowserMsg.requiredTabId();
