+++ extension/chrome/settings/setup.ts
@@ -4,7 +4,7 @@
 
 import { Store } from '../../js/common/platform/store.js';
 import { Value } from '../../js/common/core/common.js';
-import { Xss, Ui, KeyImportUi, UserAlert, KeyCanBeFixed, Env } from '../../js/common/browser.js';
+import { Xss, Ui, Env } from '../../js/common/browser.js';
 import { BrowserMsg, Bm } from '../../js/common/extension.js';
 import { Rules } from '../../js/common/rules.js';
 import { Lang } from '../../js/common/lang.js';
@@ -14,6 +14,8 @@ import { Pgp, Contact } from '../../js/common/core/pgp.js';
 import { Catch } from '../../js/common/platform/catch.js';
 import { Google, GoogleAuth } from '../../js/common/api/google.js';
 import { Attester } from '../../js/common/api/attester.js';
+import { Assert } from '../../js/common/assert.js';
+import { KeyImportUi, UserAlert, KeyCanBeFixed } from '../../js/common/ui/key_import_ui.js';
 
 declare const openpgp: typeof OpenPGP;
 
@@ -31,11 +33,11 @@ interface SetupOptions {
 Catch.try(async () => {
 
   const uncheckedUrlParams = Env.urlParams(['acctEmail', 'action', 'parentTabId']);
-  const acctEmail = Env.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
+  const acctEmail = Assert.urlParamRequire.string(uncheckedUrlParams, 'acctEmail');
   let parentTabId: string | undefined;
-  const action = Env.urlParamRequire.oneof(uncheckedUrlParams, 'action', ['add_key', 'finalize', undefined]) as 'add_key' | 'finalize' | undefined;
+  const action = Assert.urlParamRequire.oneof(uncheckedUrlParams, 'action', ['add_key', 'finalize', undefined]) as 'add_key' | 'finalize' | undefined;
   if (action === 'add_key') {
-    parentTabId = Env.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
+    parentTabId = Assert.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
   }
 
   if (acctEmail) {
@@ -275,7 +277,7 @@ Catch.try(async () => {
 
   const finalizeSetup = async ({ submit_main, submit_all }: { submit_main: boolean, submit_all: boolean }): Promise<void> => {
     const [primaryKi] = await Store.keysGet(acctEmail, ['primary']);
-    Ui.abortAndRenderErrorIfKeyinfoEmpty(primaryKi);
+    Assert.abortAndRenderErrorIfKeyinfoEmpty(primaryKi);
     try {
       await submitPublicKeyIfNeeded(primaryKi.public, { submit_main, submit_all });
     } catch (e) {
