+++ extension/js/common/browser.ts
@@ -5,36 +5,20 @@
 'use strict';
 
 import * as DOMPurify from 'dompurify';
-import { Catch, UnreportableError } from './platform/catch.js';
+import { Catch } from './platform/catch.js';
 import { Store } from './platform/store.js';
 import { Str, Value, Dict } from './core/common.js';
-import { BrowserMsg } from './extension.js';
 import { Api } from './api/api.js';
-import { Pgp, Pwd, PgpMsg, KeyInfo } from './core/pgp.js';
-import { mnemonic } from './core/mnemonic.js';
 import { Att } from './core/att.js';
-import { MsgBlock, KeyBlockType } from './core/mime.js';
-import { Settings } from './settings.js';
 import Swal from 'sweetalert2';
 import { requireTag } from './platform/require.js';
 
 requireTag('sweetalert2.js', 'sweetalert2.css');
 
-declare const openpgp: typeof OpenPGP;
-declare const qq: any;
-
-type Placement = 'settings' | 'settings_compose' | 'default' | 'dialog' | 'gmail' | 'embedded' | 'compose';
 type PreventableEventName = 'double' | 'parallel' | 'spree' | 'slowspree' | 'veryslowspree';
 type NamedSels = Dict<JQuery<HTMLElement>>;
-type KeyImportUiCheckResult = {
-  normalized: string; longid: string; passphrase: string; fingerprint: string; decrypted: OpenPGP.key.Key;
-  encrypted: OpenPGP.key.Key;
-};
 
-export type AttLimits = { count?: number, size?: number, sizeMb?: number, oversize?: (newFileSize: number) => Promise<void> };
 export type WebMailName = 'gmail' | 'outlook' | 'settings';
-export type WebmailVariantString = undefined | 'html' | 'standard' | 'new';
-export type PassphraseDialogType = 'embedded' | 'message' | 'attachment' | 'draft' | 'sign';
 export type BrowserEventErrHandler = { auth?: () => Promise<void>, authPopup?: () => Promise<void>, network?: () => Promise<void>, other?: (e: any) => Promise<void> };
 export type SelCache = { cached: (name: string) => JQuery<HTMLElement>; now: (name: string) => JQuery<HTMLElement>; sel: (name: string) => string; };
 export type UrlParam = string | number | null | undefined | boolean | string[];
@@ -127,22 +111,6 @@ export class Env {
   public static isBackgroundPage = () => Boolean(window.location && window.location.href.includes('background_page.htm'));
 
   public static isExtension = () => typeof Env.runtimeId() !== 'undefined';
-
-  public static urlParamRequire = {
-    string: (values: UrlParams, name: string): string => {
-      return String(Ui.abortAndRenderErrOnUrlParamTypeMismatch(values, name, 'string'));
-    },
-    optionalString: (values: UrlParams, name: string): string | undefined => {
-      const r = Ui.abortAndRenderErrOnUrlParamTypeMismatch(values, name, 'string?');
-      if (typeof r === 'string' || typeof r === 'undefined') {
-        return r;
-      }
-      throw new Error(`urlParamRequire.optionalString: type of ${name} unexpectedly ${typeof r}`);
-    },
-    oneof: <T>(values: UrlParams, name: string, allowed: T[]): T => {
-      return Ui.abortAndRenderErrOnUrlParamValMismatch(values, name, allowed as any as UrlParam[]) as any as T; // todo - there should be a better way
-    },
-  };
 
   private static snakeCaseToCamelCase = (s: string) => s.replace(/_[a-z]/g, boundary => boundary[1].toUpperCase());
 
@@ -272,65 +240,6 @@ export class Ui {
         }));
       }
     });
-  }
-
-  static abortAndRenderErrorIfKeyinfoEmpty = (ki: KeyInfo | undefined, doThrow: boolean = true) => {
-    if (!ki) {
-      const msg = `Cannot find primary key. Is FlowCrypt not set up yet? ${Ui.retryLink()}`;
-      Xss.sanitizeRender($('#content').length ? '#content' : 'body', msg);
-      if (doThrow) {
-        throw new UnreportableError(msg);
-      }
-    }
-  }
-
-  public static abortAndRenderErrOnUnprotectedKey = async (acctEmail?: string, tabId?: string) => {
-    if (acctEmail) {
-      const [primaryKi] = await Store.keysGet(acctEmail, ['primary']);
-      const { setup_done, setup_simple } = await Store.getAcct(acctEmail, ['setup_simple', 'setup_done']);
-      if (setup_done && setup_simple && primaryKi && (await openpgp.key.readArmored(primaryKi.private)).keys[0].isDecrypted()) {
-        if (window.location.pathname === '/chrome/settings/index.htm') {
-          // @ts-ignore - this lets it compile in content script that is missing Settings
-          Settings.renderSubPage(acctEmail, tabId!, '/chrome/settings/modules/change_passphrase.htm');
-        } else {
-          const msg = `Protect your key with a pass phrase to finish setup.`;
-          const r = await Ui.renderOverlayPromptAwaitUserChoice({ finishSetup: {}, later: { color: 'gray' } }, msg);
-          if (r === 'finish_setup') {
-            BrowserMsg.send.bg.settings({ acctEmail });
-          }
-        }
-      }
-    }
-  }
-
-  public static abortAndRenderErrOnUrlParamTypeMismatch = (values: UrlParams, name: string, expectedType: string): UrlParam => {
-    const actualType = values[name] === null ? 'null' : typeof values[name];
-    if (actualType === expectedType.replace(/\?$/, '')) { // eg expected string or optional string, and got string
-      return values[name];
-    }
-    if (actualType === 'undefined' && expectedType.match(/\?$/)) { // optional type, got undefined: ok
-      return values[name];
-    }
-    console.info(values[name]);  // for local debugging
-    // tslint:disable-next-line:max-line-length
-    const msg = `Cannot render page (expected ${Xss.escape(name)} to be of type ${Xss.escape(expectedType)} but got ${Xss.escape(actualType)})`;
-    const renderMsg = `${msg}<br><br><div class="button green long action_report_issue">report issue</div>`;
-    Xss.sanitizeRender('body', renderMsg).addClass('bad').css({ padding: '20px', 'font-size': '16px' });
-    $('.action_report_issue').click(Ui.event.handle(async target => {
-      Catch.report(msg, { currentUrl: window.location.href, params: values });
-      $('body').text('Thank you. Feel free to reach out to human@flowcrypt.com in you need assistance.');
-    }));
-    throw new UnreportableError(msg);
-  }
-
-  public static abortAndRenderErrOnUrlParamValMismatch = <T>(values: Dict<T>, name: string, expectedVals: T[]): T => {
-    if (expectedVals.indexOf(values[name]) === -1) {
-      // tslint:disable-next-line:max-line-length
-      const msg = `Cannot render page (expected ${Xss.escape(name)} to be one of ${Xss.escape(expectedVals.map(String).join(','))} but got ${Xss.escape(String(values[name]))}<br><br>Was the URL editted manually? Please write human@flowcrypt.com for help.`;
-      Xss.sanitizeRender('body', msg).addClass('bad').css({ padding: '20px', 'font-size': '16px' });
-      throw new UnreportableError(msg);
-    }
-    return values[name];
   }
 
   public static passphraseToggle = async (passphraseInputIds: string[], forceInitialShowOrHide?: "show" | "hide") => {
@@ -514,53 +423,6 @@ export class Ui {
     }
   };
 
-  /**
-   * XSS WARNING
-   *
-   * Return values are inserted directly into DOM. Results must be html escaped.
-   *
-   * When edited, REQUEST A SECOND SET OF EYES TO REVIEW CHANGES
-   */
-  public static renderableMsgBlock = (factory: XssSafeFactory, block: MsgBlock, msgId?: string, senderEmail?: string, isOutgoing?: boolean) => {
-    if (block.type === 'plainText' || block.type === 'privateKey') {
-      return Xss.escape(block.content.toString()).replace(/\n/g, '<br>') + '<br><br>';
-    } else if (block.type === 'plainHtml') {
-      return Xss.htmlSanitizeAndStripAllTags(block.content.toString(), '<br>') + '<br><br>';
-    } else if (block.type === 'encryptedMsg') {
-      return factory.embeddedMsg(block.complete ? Pgp.armor.normalize(block.content.toString(), 'encryptedMsg') : '', msgId, isOutgoing, senderEmail, false);
-    } else if (block.type === 'signedMsg') {
-      return factory.embeddedMsg(block.content.toString(), msgId, isOutgoing, senderEmail, false);
-    } else if (block.type === 'publicKey') {
-      return factory.embeddedPubkey(Pgp.armor.normalize(block.content.toString(), 'publicKey'), isOutgoing);
-    } else if (block.type === 'encryptedMsgLink') {
-      return factory.embeddedMsg('', msgId, isOutgoing, senderEmail, true, undefined, block.content.toString()); // here block.content is message short id
-    } else if (block.type === 'cryptupVerification') {
-      return factory.embeddedVerification(block.content.toString());
-    } else {
-      Catch.report(`dunno how to process block type: ${block.type}`);
-      return '';
-    }
-  }
-
-  /**
-   * XSS WARNING
-   *
-   * Return values are inserted directly into DOM. Results must be html escaped.
-   *
-   * When edited, REQUEST A SECOND SET OF EYES TO REVIEW CHANGES
-   */
-  public static replaceRenderableMsgBlocks = (factory: XssSafeFactory, origText: string, msgId?: string, senderEmail?: string, isOutgoing?: boolean) => {
-    const { blocks } = Pgp.armor.detectBlocks(origText);
-    if (blocks.length === 1 && blocks[0].type === 'plainText') {
-      return undefined; // only has single block which is plain text - meaning
-    }
-    let r = '';
-    for (const block of blocks) {
-      r += (r ? '\n\n' : '') + Ui.renderableMsgBlock(factory, block, msgId, senderEmail, isOutgoing);
-    }
-    return r;
-  }
-
   public static time = {
     wait: (untilThisFunctionEvalsTrue: () => boolean | undefined) => new Promise((success, error) => {
       const interval = Catch.setHandledInterval(() => {
@@ -736,642 +598,3 @@ export class Xss {
   }
 
 }
-
-export type FactoryReplyParams = {
-  threadId?: string,
-  threadMsgId?: string,
-  addresses?: string[],
-  replyTo?: string[],
-  myEmail?: string,
-  subject?: string,
-};
-
-export class XssSafeFactory {
-
-  /**
-   * XSS WARNING
-   *
-   * Method return values are inserted directly into DOM.
-   *
-   * All public methods are expected to escape unknown content to prevent XSS.
-   *
-   * If you add or edit a method, REQUEST A SECOND SET OF EYES TO REVIEW CHANGES
-   */
-
-  private setParams: UrlParams;
-  private reloadableCls: string;
-  private destroyableCls: string;
-  private acctEmail: string;
-  private hideGmailNewMsgInThreadNotification = '<style>.ata-asE { display: none !important; visibility: hidden !important; }</style>';
-
-  constructor(acctEmail: string, parentTabId: string, reloadableCls: string = '', destroyableCls: string = '', setParams: UrlParams = {}) {
-    this.reloadableCls = Xss.escape(reloadableCls);
-    this.destroyableCls = Xss.escape(destroyableCls);
-    this.setParams = setParams;
-    this.setParams.acctEmail = acctEmail;
-    this.setParams.parentTabId = parentTabId;
-    this.acctEmail = acctEmail;
-  }
-
-  srcImg = (relPath: string) => this.extUrl(`img/${relPath}`);
-
-  private frameSrc = (path: string, params: UrlParams = {}) => {
-    for (const k of Object.keys(this.setParams)) {
-      params[k] = this.setParams[k];
-    }
-    return Env.urlCreate(path, params);
-  }
-
-  srcComposeMsg = (draftId?: string) => {
-    return this.frameSrc(this.extUrl('chrome/elements/compose.htm'), { frameId: this.newId(), isReplyBox: false, draftId, placement: 'gmail' });
-  }
-
-  srcPassphraseDialog = (longids: string[] = [], type: PassphraseDialogType) => {
-    return this.frameSrc(this.extUrl('chrome/elements/passphrase.htm'), { type, longids });
-  }
-
-  srcSubscribeDialog = (verificationEmailText?: string, placement?: Placement, isAuthErr?: boolean) => {
-    return this.frameSrc(this.extUrl('chrome/elements/subscribe.htm'), { verificationEmailText, placement, isAuthErr });
-  }
-
-  srcVerificationDialog = (verificationEmailText: string) => {
-    return this.frameSrc(this.extUrl('chrome/elements/verification.htm'), { verificationEmailText });
-  }
-
-  srcAddPubkeyDialog = (emails: string[], placement: Placement) => {
-    return this.frameSrc(this.extUrl('chrome/elements/add_pubkey.htm'), { emails, placement });
-  }
-
-  srcAddFooterDialog = (placement: Placement, grandparentTabId: string) => {
-    return this.frameSrc(this.extUrl('chrome/elements/shared/footer.htm'), { placement, grandparentTabId });
-  }
-
-  srcSendingAddrDialog = (placement: Placement) => {
-    return this.frameSrc(this.extUrl('chrome/elements/sending_address.htm'), { placement });
-  }
-
-  srcPgpAttIframe = (a: Att, isEncrypted: boolean) => {
-    if (!a.id && !a.url && a.hasData()) { // data provided directly, pass as object url
-      a.url = Browser.objUrlCreate(a.getData());
-    }
-    return this.frameSrc(this.extUrl('chrome/elements/attachment.htm'), {
-      frameId: this.newId(), msgId: a.msgId, name: a.name, type: a.type, size: a.length, attId: a.id, url: a.url, isEncrypted
-    });
-  }
-
-  srcPgpBlockIframe = (message: string, msgId?: string, isOutgoing?: boolean, senderEmail?: string, hasPassword?: boolean, signature?: string | boolean, short?: string) => {
-    return this.frameSrc(this.extUrl('chrome/elements/pgp_block.htm'), { frameId: this.newId(), message, hasPassword, msgId, senderEmail, isOutgoing, signature, short });
-  }
-
-  srcPgpPubkeyIframe = (armoredPubkey: string, isOutgoing?: boolean) => {
-    return this.frameSrc(this.extUrl('chrome/elements/pgp_pubkey.htm'), { frameId: this.newId(), armoredPubkey, minimized: Boolean(isOutgoing), });
-  }
-
-  srcBackupIframe = (armoredPubkey: string) => {
-    return this.frameSrc(this.extUrl('chrome/elements/backup.htm'), { frameId: this.newId(), armoredPubkey });
-  }
-
-  srcReplyMsgIframe = (convoParams: FactoryReplyParams, skipClickPrompt: boolean, ignoreDraft: boolean) => {
-    const params: UrlParams = {
-      isReplyBox: true,
-      frameId: `frame_${Str.sloppyRandom(10)}`,
-      placement: 'gmail',
-      threadId: convoParams.threadId,
-      skipClickPrompt: Boolean(skipClickPrompt),
-      ignoreDraft: Boolean(ignoreDraft),
-      threadMsgId: convoParams.threadMsgId,
-    };
-    if (convoParams.replyTo) { // for gmail and inbox. Outlook gets this from API
-      const headers = this.resolveFromTo(convoParams.addresses || [], convoParams.myEmail || this.acctEmail, convoParams.replyTo);
-      params.to = headers.to;
-      params.from = headers.from;
-      params.subject = convoParams.subject ? 'Re: ' + convoParams.subject : '';
-    }
-    return this.frameSrc(this.extUrl('chrome/elements/compose.htm'), params);
-  }
-
-  srcStripeCheckout = () => {
-    return this.frameSrc('https://flowcrypt.com/stripe.htm', {});
-  }
-
-  metaNotificationContainer = () => {
-    return `<div class="${this.destroyableCls} webmail_notifications" style="text-align: center;"></div>`;
-  }
-
-  metaStylesheet = (file: string) => {
-    return `<link class="${this.destroyableCls}" rel="stylesheet" href="${this.extUrl(`css/${file}.css`)}" />`;
-  }
-
-  dialogPassphrase = (longids: string[], type: PassphraseDialogType) => {
-    return this.divDialog_DANGEROUS(this.iframe(this.srcPassphraseDialog(longids, type), ['medium'], { scrolling: 'no' }), 'dialog-passphrase'); // xss-safe-factory
-  }
-
-  dialogSubscribe = (verifEmailText?: string, isAuthErr?: boolean) => {
-    const src = this.srcSubscribeDialog(verifEmailText, 'dialog', isAuthErr);
-    return this.divDialog_DANGEROUS(this.iframe(src, ['mediumtall'], { scrolling: 'no' }), 'dialog-subscribe'); // xss-safe-factory
-  }
-
-  dialogAddPubkey = (emails: string[]) => {
-    return this.divDialog_DANGEROUS(this.iframe(this.srcAddPubkeyDialog(emails, 'gmail'), ['tall'], { scrolling: 'no' }), 'dialog-add-pubkey'); // xss-safe-factory
-  }
-
-  embeddedCompose = (draftId?: string) => {
-    return Ui.e('div', { id: 'new_message', class: 'new_message', 'data-test': 'container-new-message', html: this.iframe(this.srcComposeMsg(draftId), [], { scrolling: 'no' }) });
-  }
-
-  embeddedSubscribe = (verifEmailText: string, isAuthErr: boolean) => {
-    return this.iframe(this.srcSubscribeDialog(verifEmailText, 'embedded', isAuthErr), ['short', 'embedded'], { scrolling: 'no' });
-  }
-
-  embeddedVerification = (verifEmailText: string) => {
-    return this.iframe(this.srcVerificationDialog(verifEmailText), ['short', 'embedded'], { scrolling: 'no' });
-  }
-
-  embeddedAtta = (meta: Att, isEncrypted: boolean) => {
-    return Ui.e('span', { class: 'pgp_attachment', html: this.iframe(this.srcPgpAttIframe(meta, isEncrypted)) });
-  }
-
-  embeddedMsg = (armored: string, msgId?: string, isOutgoing?: boolean, sender?: string, hasPassword?: boolean, signature?: string | boolean, short?: string) => {
-    return this.iframe(this.srcPgpBlockIframe(armored, msgId, isOutgoing, sender, hasPassword, signature, short), ['pgp_block']) + this.hideGmailNewMsgInThreadNotification;
-  }
-
-  embeddedPubkey = (armoredPubkey: string, isOutgoing?: boolean) => {
-    return this.iframe(this.srcPgpPubkeyIframe(armoredPubkey, isOutgoing), ['pgp_block']);
-  }
-
-  embeddedBackup = (armoredPubkey: string) => {
-    return this.iframe(this.srcBackupIframe(armoredPubkey), ['backup_block']);
-  }
-
-  embeddedReply = (convoParams: FactoryReplyParams, skipClickPrompt: boolean, ignoreDraft: boolean = false) => {
-    return this.iframe(this.srcReplyMsgIframe(convoParams, skipClickPrompt, ignoreDraft), ['reply_message']);
-  }
-
-  embeddedPassphrase = (longids: string[]) => {
-    return this.divDialog_DANGEROUS(this.iframe(this.srcPassphraseDialog(longids, 'embedded'), ['medium'], { scrolling: 'no' }), 'embedded-passphrase'); // xss-safe-factory
-  }
-
-  embeddedAttaStatus = (content: string) => {
-    return Ui.e('div', { class: 'attachment_loader', html: Xss.htmlSanitize(content) });
-  }
-
-  embeddedStripeCheckout = () => {
-    return this.iframe(this.srcStripeCheckout(), [], { sandbox: 'allow-forms allow-scripts allow-same-origin' });
-  }
-
-  btnCompose = (webmailName: WebMailName) => {
-    if (webmailName === 'outlook') {
-      const btn = `<div class="new_message_button" title="New Secure Email"><img src="${this.srcImg('logo-19-19.png')}"></div>`;
-      return `<div class="_fce_c ${this.destroyableCls} cryptup_compose_button_container" role="presentation">${btn}</div>`;
-    } else {
-      const btn = `<div class="new_message_button T-I J-J5-Ji T-I-KE L3" id="flowcrypt_new_message_button" role="button" tabindex="0" data-test="action-secure-compose">Secure Compose</div>`;
-      return `<div class="${this.destroyableCls} z0">${btn}</div>`;
-    }
-  }
-
-  btnReply = () => {
-    return `<div class="${this.destroyableCls} reply_message_button"><img src="${this.srcImg('svgs/reply-icon.svg')}" /></div>`;
-  }
-
-  btnWithoutFc = () => {
-    const span = `<span>see original</span>`;
-    return `<span class="hk J-J5-Ji cryptup_convo_button show_original_conversation ${this.destroyableCls}" data-tooltip="Show conversation without FlowCrypt">${span}</span>`;
-  }
-
-  btnWithFc = () => {
-    return `<span class="hk J-J5-Ji cryptup_convo_button use_secure_reply ${this.destroyableCls}" data-tooltip="Use Secure Reply"><span>secure reply</span></span>`;
-  }
-
-  btnRecipientsUseEncryption = (webmailName: WebMailName) => {
-    if (webmailName !== 'gmail') {
-      Catch.report('switch_to_secure not implemented for ' + webmailName);
-      return '';
-    } else {
-      return '<div class="aoD az6 recipients_use_encryption">Your recipients seem to have encryption set up! <a href="#">Secure Compose</a></div>';
-    }
-  }
-
-  btnSettings = (webmailName: WebMailName) => {
-    if (webmailName !== 'gmail') {
-      Catch.report('btnSettings not implemented for ' + webmailName);
-      return '';
-    } else {
-      return `<div id="fc_settings_btn" class="f1">FlowCrypt</div>`;
-    }
-  }
-
-  private extUrl = (s: string) => chrome.runtime.getURL(s);
-
-  private newId = () => `frame_${Str.sloppyRandom(10)}`;
-
-  private resolveFromTo = (secondaryEmails: string[], myEmail: string, theirEmails: string[]) => {
-    // when replaying to email I've sent myself, make sure to send it to the other person, and not myself
-    if (theirEmails.length === 1 && secondaryEmails.includes(theirEmails[0])) {
-      return { from: theirEmails[0], to: myEmail }; // replying to myself, reverse the values to actually write to them
-    }
-    return { to: theirEmails, from: myEmail };
-  }
-
-  private iframe = (src: string, classes: string[] = [], elAttributes: UrlParams = {}) => {
-    const id = String(Env.urlParams(['frameId'], src).frameId);
-    const classAttribute = (classes || []).concat(this.reloadableCls).join(' ');
-    const attrs: Dict<string> = { id, class: classAttribute, src };
-    for (const name of Object.keys(elAttributes)) {
-      attrs[name] = String(elAttributes[name]);
-    }
-    return Ui.e('iframe', attrs);
-  }
-
-  // tslint:disable-next-line:variable-name
-  private divDialog_DANGEROUS = (content_MUST_BE_XSS_SAFE: string, dataTest: string) => { // xss-dangerous-function
-    return Ui.e('div', { id: 'cryptup_dialog', html: content_MUST_BE_XSS_SAFE, 'data-test': dataTest });
-  }
-
-}
-
-export class KeyCanBeFixed extends Error {
-  encrypted: OpenPGP.key.Key;
-  constructor(encrypted: OpenPGP.key.Key) {
-    super();
-    this.encrypted = encrypted;
-  }
-}
-
-export class UserAlert extends Error { }
-
-export class KeyImportUi {
-
-  private expectedLongid?: string;
-  private rejectKnown: boolean;
-  private checkEncryption: boolean;
-  private checkSigning: boolean;
-  public onBadPassphrase: VoidCallback = () => undefined;
-
-  constructor(o: { expectLongid?: string, rejectKnown?: boolean, checkEncryption?: boolean, checkSigning?: boolean }) {
-    this.expectedLongid = o.expectLongid;
-    this.rejectKnown = o.rejectKnown === true;
-    this.checkEncryption = o.checkEncryption === true;
-    this.checkSigning = o.checkSigning === true;
-  }
-
-  public initPrvImportSrcForm = (acctEmail: string, parentTabId: string | undefined) => {
-    $('input[type=radio][name=source]').off().change(function () {
-      if ((this as HTMLInputElement).value === 'file') {
-        $('.input_private_key').val('').change().prop('disabled', true);
-        $('.source_paste_container').css('display', 'none');
-        $('.source_paste_container .pass_phrase_needed').hide();
-        $('#fineuploader_button > input').click();
-      } else if ((this as HTMLInputElement).value === 'paste') {
-        $('.input_private_key').val('').change().prop('disabled', false);
-        $('.source_paste_container').css('display', 'block');
-        $('.source_paste_container .pass_phrase_needed').hide();
-      } else if ((this as HTMLInputElement).value === 'backup') {
-        window.location.href = Env.urlCreate('/chrome/settings/setup.htm', { acctEmail, parentTabId, action: 'add_key' });
-      }
-    });
-    $('.line.pass_phrase_needed .action_use_random_pass_phrase').click(Ui.event.handle(target => {
-      $('.source_paste_container .input_passphrase').val(Pgp.password.random());
-      $('.input_passphrase').attr('type', 'text');
-      $('#e_rememberPassphrase').prop('checked', true);
-    }));
-    $('.input_private_key').change(Ui.event.handle(async target => {
-      const { keys: [prv] } = await openpgp.key.readArmored(String($(target).val()));
-      $('.input_passphrase').val('');
-      if (prv && prv.isPrivate() && prv.isDecrypted()) {
-        $('.line.pass_phrase_needed').show();
-      } else {
-        $('.line.pass_phrase_needed').hide();
-      }
-    }));
-    const attach = new AttUI(() => Promise.resolve({ count: 100, size: 1024 * 1024, size_mb: 1 }));
-    attach.initAttDialog('fineuploader', 'fineuploader_button');
-    attach.setAttAddedCb(async file => {
-      let prv: OpenPGP.key.Key | undefined;
-      const utf = file.getData().toUtfStr();
-      if (utf.includes(Pgp.armor.headers('privateKey').begin)) {
-        const firstPrv = Pgp.armor.detectBlocks(utf).blocks.filter(b => b.type === 'privateKey')[0];
-        if (firstPrv) { // filter out all content except for the first encountered private key (GPGKeychain compatibility)
-          prv = (await openpgp.key.readArmored(firstPrv.content.toString())).keys[0];
-        }
-      } else {
-        prv = (await openpgp.key.read(file.getData())).keys[0];
-      }
-      if (typeof prv !== 'undefined') {
-        $('.input_private_key').val(prv.armor()).change().prop('disabled', true);
-        $('.source_paste_container').css('display', 'block');
-      } else {
-        $('.input_private_key').val('').change().prop('disabled', false);
-        await Ui.modal.error('Not able to read this key. Is it a valid PGP private key?');
-        $('input[type=radio][name=source]').removeAttr('checked');
-      }
-    });
-  }
-
-  checkPrv = async (acctEmail: string, armored: string, passphrase: string): Promise<KeyImportUiCheckResult> => {
-    const { normalized } = await this.normalize('privateKey', armored);
-    const decrypted = await this.read('privateKey', normalized);
-    const encrypted = await this.read('privateKey', normalized);
-    const longid = await this.longid(decrypted);
-    this.rejectIfNot('privateKey', decrypted);
-    await this.rejectKnownIfSelected(acctEmail, decrypted);
-    this.rejectIfDifferentFromSelectedLongid(longid);
-    await this.decryptAndEncryptAsNeeded(decrypted, encrypted, passphrase);
-    await this.checkEncryptionPrvIfSelected(decrypted, encrypted);
-    await this.checkSigningIfSelected(decrypted);
-    return { normalized, longid, passphrase, fingerprint: (await Pgp.key.fingerprint(decrypted))!, decrypted, encrypted }; // will have fp if had longid
-  }
-
-  checkPub = async (armored: string): Promise<string> => {
-    const { normalized } = await this.normalize('publicKey', armored);
-    const parsed = await this.read('publicKey', normalized);
-    await this.longid(parsed);
-    await this.checkEncryptionPubIfSelected(normalized);
-    return normalized;
-  }
-
-  private normalize = async (type: KeyBlockType, armored: string) => {
-    const headers = Pgp.armor.headers(type);
-    const normalized = await Pgp.key.normalize(armored);
-    if (!normalized) {
-      throw new UserAlert('There was an error processing this key, possibly due to bad formatting.\nPlease insert complete key, including "' + headers.begin + '" and "' + headers.end + '"');
-    }
-    return normalized;
-  }
-
-  private read = async (type: KeyBlockType, normalized: string) => {
-    const headers = Pgp.armor.headers(type);
-    const { keys: [k] } = await openpgp.key.readArmored(normalized);
-    if (typeof k === 'undefined') {
-      throw new UserAlert('Private key is not correctly formated. Please insert complete key, including "' + headers.begin + '" and "' + headers.end + '"');
-    }
-    return k;
-  }
-
-  private longid = async (k: OpenPGP.key.Key) => {
-    const longid = await Pgp.key.longid(k);
-    if (!longid) {
-      throw new UserAlert('This key may not be compatible. Email human@flowcrypt.com and const us know which software created this key.\n\n(error: cannot get long_id)');
-    }
-    return longid;
-  }
-
-  private rejectIfNot = (type: KeyBlockType, k: OpenPGP.key.Key) => {
-    const headers = Pgp.armor.headers(type);
-    if (type === 'privateKey' && k.isPublic()) {
-      throw new UserAlert('This was a public key. Please insert a private key instead. It\'s a block of text starting with "' + headers.begin + '"');
-    }
-    if (type === 'publicKey' && !k.isPublic()) {
-      throw new UserAlert('This was a public key. Please insert a private key instead. It\'s a block of text starting with "' + headers.begin + '"');
-    }
-  }
-
-  private rejectKnownIfSelected = async (acctEmail: string, k: OpenPGP.key.Key) => {
-    if (this.rejectKnown) {
-      const keyinfos = await Store.keysGet(acctEmail);
-      const privateKeysLongids = keyinfos.map(ki => ki.longid);
-      if (privateKeysLongids.includes(String(await Pgp.key.longid(k)))) {
-        throw new UserAlert('This is one of your current keys, try another one.');
-      }
-    }
-  }
-
-  private rejectIfDifferentFromSelectedLongid = (longid: string) => {
-    if (this.expectedLongid && longid !== this.expectedLongid) {
-      throw new UserAlert(`Key does not match. Looking for key with KeyWords ${mnemonic(this.expectedLongid)} (${this.expectedLongid})`);
-    }
-  }
-
-  private decryptAndEncryptAsNeeded = async (toDecrypt: OpenPGP.key.Key, toEncrypt: OpenPGP.key.Key, passphrase: string): Promise<void> => {
-    if (!passphrase) {
-      throw new UserAlert('Please enter a pass phrase to use with this key');
-    }
-    let decryptResult;
-    try {
-      if (toEncrypt.isDecrypted()) {
-        await toEncrypt.encrypt(passphrase);
-      }
-      if (toDecrypt.isDecrypted()) {
-        return;
-      }
-      decryptResult = await Pgp.key.decrypt(toDecrypt, [passphrase]);
-    } catch (e) {
-      throw new UserAlert(`This key is not supported by FlowCrypt yet. Please write at human@flowcrypt.com to add support soon. (decrypt error: ${String(e)})`);
-    }
-    if (!decryptResult) {
-      this.onBadPassphrase();
-      if (this.expectedLongid) {
-        // tslint:disable-next-line:max-line-length
-        throw new UserAlert('This is the right key! However, the pass phrase does not match. Please try a different pass phrase. Your original pass phrase might have been different then what you use now.');
-      } else {
-        throw new UserAlert('The pass phrase does not match. Please try a different pass phrase.');
-      }
-    }
-  }
-
-  private checkEncryptionPrvIfSelected = async (k: OpenPGP.key.Key, encrypted: OpenPGP.key.Key) => {
-    if (this.checkEncryption && ! await k.getEncryptionKey()) {
-      if (await k.verifyPrimaryKey() === openpgp.enums.keyStatus.no_self_cert || await Pgp.key.usableButExpired(k)) { // known issues - key can be fixed
-        throw new KeyCanBeFixed(encrypted);
-      } else {
-        throw new UserAlert('This looks like a valid key but it cannot be used for encryption. Please write at human@flowcrypt.com to see why is that.');
-      }
-    }
-  }
-
-  private checkEncryptionPubIfSelected = async (normalized: string) => {
-    if (this.checkEncryption && !await Pgp.key.usable(normalized)) {
-      throw new UserAlert('This public key looks correctly formatted, but cannot be used for encryption. Please write at human@flowcrypt.com. We\'ll see if there is a way to fix it.');
-    }
-  }
-
-  private checkSigningIfSelected = async (k: OpenPGP.key.Key) => {
-    if (this.checkSigning && ! await k.getSigningKey()) {
-      throw new UserAlert('This looks like a valid key but it cannot be used for signing. Please write at human@flowcrypt.com to see why is that.');
-    }
-  }
-}
-
-export class AttUI {
-
-  private templatePath = '/chrome/elements/shared/attach.template.htm';
-  private getLimits: () => Promise<AttLimits>;
-  private attachedFiles: Dict<File> = {};
-  private uploader: any = undefined;
-  private attAddedCb?: (r: Att) => Promise<void>;
-  private attRemovedCb?: () => void;
-
-  constructor(getLimits: () => Promise<AttLimits>) {
-    this.getLimits = getLimits;
-  }
-
-  initAttDialog = (elId: string, btnId: string) => {
-    $('#qq-template').load(this.templatePath, () => {
-      const config = {
-        autoUpload: false,
-        // debug: true,
-        element: $('#' + elId).get(0),
-        button: $('#' + btnId).get(0),
-        dragAndDrop: {
-          extraDropzones: $('#input_text'),
-        },
-        callbacks: {
-          onSubmitted: (uploadFileId: string, name: string) => this.processNewAtt(uploadFileId, name).catch(Catch.reportErr),
-          onCancel: (uploadFileId: string) => Catch.try(() => this.cancelAtt(uploadFileId))(),
-        },
-      };
-      this.uploader = new qq.FineUploader(config); // tslint:disable-line:no-unsafe-any
-    });
-  }
-
-  setAttAddedCb = (cb: (r: Att) => Promise<void>) => {
-    this.attAddedCb = cb;
-  }
-
-  setAttRemovedCb = (cb: () => void) => {
-    this.attRemovedCb = cb;
-  }
-
-  hasAtt = () => {
-    return Object.keys(this.attachedFiles).length > 0;
-  }
-
-  getAttIds = () => {
-    return Object.keys(this.attachedFiles);
-  }
-
-  collectAtt = async (uploadFileId: string) => {
-    const fileData = await this.readAttDataAsUint8(uploadFileId);
-    return new Att({ name: this.attachedFiles[uploadFileId].name, type: this.attachedFiles[uploadFileId].type, data: fileData });
-  }
-
-  collectAtts = async () => {
-    const atts: Att[] = [];
-    for (const uploadFileId of Object.keys(this.attachedFiles)) {
-      atts.push(await this.collectAtt(uploadFileId));
-    }
-    return atts;
-  }
-
-  collectEncryptAtts = async (pubkeys: string[], pwd?: Pwd): Promise<Att[]> => {
-    const atts: Att[] = [];
-    for (const uploadFileId of Object.keys(this.attachedFiles)) {
-      const file = this.attachedFiles[uploadFileId];
-      const data = await this.readAttDataAsUint8(uploadFileId);
-      const encrypted = await PgpMsg.encrypt({ pubkeys, data, pwd, filename: file.name, armor: false }) as OpenPGP.EncryptBinaryResult;
-      atts.push(new Att({ name: file.name.replace(/[^a-zA-Z\-_.0-9]/g, '_').replace(/__+/g, '_') + '.pgp', type: file.type, data: encrypted.message.packets.write() }));
-    }
-    return atts;
-  }
-
-  clearAllAtts = () => {
-    this.attachedFiles = {};
-  }
-
-  private cancelAtt = (uploadFileId: string) => {
-    delete this.attachedFiles[uploadFileId];
-    if (this.attRemovedCb) {
-      // run at next event loop cycle - let DOM changes render first
-      // this allows code that relies on this to evaluate the DOM after the file has been removed from it
-      Catch.setHandledTimeout(this.attRemovedCb, 0);
-    }
-  }
-
-  private processNewAtt = async (uploadFileId: string, name: string) => {
-    const limits = await this.getLimits();
-    if (limits.count && Object.keys(this.attachedFiles).length >= limits.count) {
-      await Ui.modal.warning('Amount of attached files is limited to ' + limits.count);
-      this.uploader.cancel(uploadFileId); // tslint:disable-line:no-unsafe-any
-    } else {
-      const newFile: File = this.uploader.getFile(uploadFileId); // tslint:disable-line:no-unsafe-any
-      if (limits.size && this.getFileSizeSum() + newFile.size > limits.size) {
-        this.uploader.cancel(uploadFileId); // tslint:disable-line:no-unsafe-any
-        if (typeof limits.oversize === 'function') {
-          await limits.oversize(this.getFileSizeSum() + newFile.size);
-        } else {
-          await Ui.modal.warning('Combined file size is limited to ' + limits.sizeMb + 'MB');
-        }
-        return;
-      }
-      this.attachedFiles[uploadFileId] = newFile;
-      if (typeof this.attAddedCb === 'function') {
-        const a = await this.collectAtt(uploadFileId);
-        await this.attAddedCb(a);
-      }
-    }
-  }
-
-  private getFileSizeSum = () => {
-    let sum = 0;
-    for (const file of Object.values(this.attachedFiles)) {
-      sum += file.size;
-    }
-    return sum;
-  }
-
-  private readAttDataAsUint8 = (uploadFileId: string): Promise<Uint8Array> => {
-    return new Promise(resolve => {
-      const reader = new FileReader();
-      reader.onload = () => {
-        resolve(new Uint8Array(reader.result as ArrayBuffer)); // that's what we're getting
-      };
-      reader.readAsArrayBuffer(this.attachedFiles[uploadFileId]);
-    });
-  }
-
-}
-
-/*
- * This function uses fineuploader to import a file of PGP public keys (binary or armored). Usage examples:
- * 'extension/chrome/settings/modules/contacts.{ts,htm}' and 'extension/chrome/elements/add_pubkey.{ts,htm}'
- *
- * In the markup you must add:
- *   1. The following elements to the documents 'head':
- *    <link rel="stylesheet" href="/css/fine-uploader-new.css" />
- *    <script type="text/template" src="/chrome/elements/hared/attach.template.htm" id="qq-template"></script>
- *   2. Find a place in the body for the following elements:
- *    <div id="fineuploader_button" class="button green long">keyring file</div>
- *    <div id="fineuploader" class="display_none"></div>
- *  3. Place this script with the other scripts near the bottom of the body:
- *    <script src="/lib/fine-uploader.js"></script>
- *
- * In the TS file you must:
- *      const attUI = new AttUI(() => Promise.resolve({ size_mb: 5, size: 5 * 1024 * 1024, count: 1 }));
- *      attUI.initAttDialog('fineuploader', 'fineuploader_button');
- *      attUI.setAttAddedCb((file) => {
- *        const keys = handleImportKeyringFile(attUI, file);
- *        // Handle keys parsed from the file
- *      });
- */
-export const processPublicKeyFileImport = async (ui: AttUI, file: Att) => {
-  ui.clearAllAtts();
-  try {
-    const { blocks } = await Pgp.armor.detectBlocks(file.getData().toUtfStr());
-    const armoredPublicKeyBlocks = blocks.filter(block => block.type === "publicKey");
-    const allKeys: OpenPGP.key.Key[] = [];
-    const allErrs: Error[] = [];
-    if (armoredPublicKeyBlocks.length) {
-      for (const block of blocks) {
-        const { err, keys } = await openpgp.key.readArmored(block.content.toString());
-        if (err) {
-          allErrs.push(...err);
-        }
-        allKeys.push(...keys);
-      }
-    } else {
-      const { err, keys } = await openpgp.key.read(file.getData());
-      if (err) {
-        allErrs.push(...err);
-      }
-      allKeys.push(...keys);
-    }
-    if (allErrs.length && !allKeys.length) {
-      await Ui.modal.error(`error processing public keys:\n${allErrs.map(e => `-> ${e.message}\n`).join('')}`);
-      return undefined;
-    }
-    return allKeys;
-  } catch (e) {
-    Catch.reportErr(e);
-    await Ui.modal.error(`error processing public keys: ${String(e)}`);
-    return undefined;
-  }
-};
