+++ extension/js/content_scripts/checkout/stripe.ts
@@ -5,11 +5,12 @@
 import { BrowserMsg } from '../../common/extension.js';
 import { Catch } from '../../common/platform/catch.js';
 import { Env } from '../../common/browser.js';
+import { Assert } from '../../common/assert.js';
 
 Catch.try(async () => {
 
   const uncheckedUrlParams = Env.urlParams(['parentTabId']);
-  const parentTabId = Env.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
+  const parentTabId = Assert.urlParamRequire.string(uncheckedUrlParams, 'parentTabId');
 
   document.addEventListener('cryptup_stripe_result', Catch.try(() => {
     BrowserMsg.send.stripeResult(parentTabId, { token: String($('#stripe_result').text()) });
